requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
langgraph==0.4.8
langchain==0.3.26
python-dotenv==1.1.1
langchain-openai
langchain-google-genai
langchain-community
faiss-cpu
pypdf
pandas
streamlit
httpx
deepeval
langchain-text-splitters
langchain-groq
langchain-huggingface
sentence-transformers
